@@ -1,12 +1,14 @@
 {
   "name": "company-ping-pong",
-  "version": "1.0.0",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "description": "Coming soon...",
-  "main": "index.js",
   "scripts": {
-    "client": "rollup --config rollup.config.js",
-    "server": "nodemon src/server/index.mjs",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "server": "nodemon src/server/index.mjs"
   },
   "repository": {
     "type": "git",
@@ -21,18 +23,18 @@
   "homepage": "https://github.com/zantinger/company-ping-pong#readme",
   "dependencies": {
     "express": "^4.17.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "milligram": "^1.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rxjs": "^7.5.1",
     "socket.io": "^4.4.0"
   },
   "devDependencies": {
-    "@babel/preset-react": "^7.16.5",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.1.1",
-    "@rollup/plugin-replace": "^3.0.1",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.0.1",
     "nodemon": "^2.0.15",
-    "rollup": "^2.62.0"
+    "socket.io-client": "^4.5.1",
+    "vite": "^3.0.7"
   }
 }
